+++ vizQA/__init__.py
@@ -1,13 +1,14 @@
 # pylint: disable=redefined-builtin
 """Top-level package exports for vizQA."""
 
-from vizQA.library import StepResult, VizQASession, attach, click, run_step, run_steps, type, verify
+from vizQA.library import StepResult, VizQASession, attach, click, key_input, run_step, run_steps, type, verify
 
 __all__ = [
     "StepResult",
     "VizQASession",
     "attach",
     "click",
+    "key_input",
     "run_step",
     "run_steps",
     "type",